% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseK.R
\name{chooseK}
\alias{chooseK}
\title{chooseK}
\usage{
chooseK(
  counts,
  pseudotimes,
  conditions,
  weights = NULL,
  experiments = NULL,
  linCompare = 1,
  n = 100,
  k_vals = 1:9,
  genes = NULL,
  parallel = FALSE,
  n_cores = NULL,
  save.name = "k_results.png",
  fit_time = 100,
  seed = 123
)
}
\arguments{
\item{counts}{matrix (size = cells by genes) of raw (integer) count values of gene expression}

\item{pseudotimes}{matrix (size = cells by lineages) of pseudotime values}

\item{conditions}{factor (length = cells) specifying condition to which each cell belongs}

\item{weights}{matrix (size = cells by lineages) of weights from pseudotime analysis if applicable (otherwise NULL)}

\item{experiments}{factor (length = cells) specifying replicate/batch to which each cell belongs if applicable (otherwise NULL)}

\item{linCompare}{integer specifying which lineage (index) to test (defaults to 1 - only one lineage provided in pseudotimes)}

\item{n}{integer number of genes to randomly sample (default = 100)}

\item{k_vals}{vector of integers specifying number of knots in NB-GAM fit to trial (default = c(1:9))}

\item{genes}{vector of gene names or indices to sample from if desired (defaults to NULL and samples from all genes)}

\item{parallel}{boolean specifying whether or not to run in parallel (requires BiocParallel)}

\item{n_cores}{integer specifying number of cores to use if running in parallel}

\item{save.name}{string specifying file name if image of AIC results to be saved (if set to NULL, no image saved). Defaults to "k_results.png".}

\item{fit_time}{time permitted (in seconds) for NB-GAM fits for single gene (defaults to 100)}

\item{seed}{set random seed}
}
\value{
a list containing AIC, R2 (r squared), and deviance explained values (matrix size = n by length(k_vals))
}
\description{
Fits alternative NB-GAM model for a random sample (n=100 by default) of genes
over different values of k (number of knots) and returns average AIC.  Minimum
AIC suggests value of k for downstream analysis.
}
