% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleInputData}
\alias{ExampleInputData}
\title{Example Input Data for LineageDE}
\usage{
data(ExampleInputData)
}
\description{
A list containing example input data to run LineageDE
}
\details{
\describe{
  \item{counts}{matrix of gene expression counts (cells x genes)}
  \item{conditions}{factor specifying experimental conditions}
  \item{experiments}{factor specifying experimental replicates}
  \item{ori.pseudotimes}{matrix of pseudotime values (cells x lineages)}
  \item{ori.weights}{matrix of pseudotime weights (cells x lineages)}
  \item{samples}{matrix of cell indices sampled (cells x number_samples=100)}
  \item{sub.pseudotimes}{matrix of sampled pseudotime values (cells x lineages x 100)}
  \item{sub.weights}{matrix of sampled pseudotime weights (cells x lineages x 100)}

}
}
