% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{plotGene}
\usage{
plotGene(
  gene,
  counts,
  pseudotimes,
  conditions,
  weights = NULL,
  experiments = NULL,
  linCompare = 1,
  k = 3,
  seed = 123
)
}
\arguments{
\item{gene}{name or index of gene to plot}

\item{counts}{matrix (size = cells by genes) of raw (integer) count values of gene expression}

\item{pseudotimes}{matrix (size = cells by lineages) of pseudotime values}

\item{conditions}{factor (length = cells) specifying condition to which each cell belongs}

\item{weights}{matrix (size = cells by lineages) of weights from pseudotime analysis if applicable (otherwise NULL)}

\item{experiments}{factor (length = cells) specifying replicate/batch to which each cell belongs if applicable (otherwise NULL)}

\item{linCompare}{integer specifying which lineage (index) to plot (defaults to 1 - only one lineage provided in pseudotimes)}

\item{k}{integer specifying number of knots in NB-GAM fit (default = 3)}

\item{seed}{set random seed}
}
\value{
ggplot object
}
\description{
Plots expression for gene of interest over pseudotime for a single lineage.
Plots both CPM and NB-GAM fits separated by specified conditions.
}
