% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGene.R
\name{testGene}
\alias{testGene}
\title{testGene}
\usage{
testGene(
  gene,
  counts,
  ori.pseudotimes,
  ori.weights,
  conditions,
  experiments,
  offsets,
  knotList,
  k,
  samples,
  sub.pseudotimes,
  sub.weights,
  fit_time,
  seed
)
}
\arguments{
\item{gene}{names or index specifying gene to be tested}

\item{counts}{matrix (size = cells by genes) of raw (integer) count values of gene expression}

\item{ori.pseudotimes}{matrix (size = cells by lineages) of pseudotime values}

\item{ori.weights}{matrix (size = cells by lineages) of weights from pseudotime analysis if applicable (otherwise NULL)}

\item{conditions}{factor (length = cells) specifying condition to which each cell belongs}

\item{experiments}{factor (length = cells) specifying replicate/batch to which each cell belongs if applicable (otherwise NULL)}

\item{offsets}{offset used in fitting NB-GAM based off of library size for each cell}

\item{knotList}{list specifying knot locations used in NB-GAM fit}

\item{k}{integer specifying number of knots in NB-GAM fit}

\item{samples}{matrix (size = cells by num_samples) specifying indices cells utilized in each sample of pseudotime if applicable (otherwise NULL)}

\item{sub.pseudotimes}{matrix (size = cells by lineages by num_samples) samples of pseudotime values if applicable (otherwise NULL)}

\item{sub.weights}{matrix (size = cells by lineages by num_samples) samples of pseudotime weights if applicable (otherwise NULL)}

\item{fit_time}{time permitted (in seconds) for NB-GAM fit for single gene}

\item{seed}{set random seed}
}
\value{
LRT, ori.pVal, (emp.pVal), (par.pVal) for gene tested
}
\description{
Fits NB-GAM to test alternative vs null model for a specific gene.  Helper function for LineageDE.
Performs likelihood ratio test and returns p-value based off of chi-squared distribution (ori.pVal)
If samples of pseudotime are provided, will also perform permutation test and give an empirical p-value (emp.pVal)
and a parametric p-value (par.pVal) from fitting null distribution with a gamma distribution
}
